Add files via upload
This repository contains a sentiment analysis model built using a Long Short-Term Memory (LSTM) neural network to classify IMDB movie reviews as positive or negative. The model is trained on a dataset of movie reviews, with the goal of understanding and predicting the sentiment expressed in the text.

Dataset
The dataset used for this project is the IMDB Movie Reviews dataset, which contains 50,000 movie reviews labeled as either positive or negative. The dataset is sourced from Kaggle and is imported directly using Kaggle's API.This repository contains a sentiment analysis model built using a Long Short-Term Memory (LSTM) neural network to classify IMDB movie reviews as positive or negative. The model is trained on a dataset of movie reviews, with the goal of understanding and predicting the sentiment expressed in the text.
Dataset:
The dataset used for this project is the IMDB Movie Reviews dataset, which contains 50,000 movie reviews labeled as either positive or negative. The dataset is sourced from Kaggle and is imported directly using Kaggle's API.